🎨 simplify db scripts
Significant amount of repetition avoided.
