+++ tests/database/queries/user_check_exists.test.js
@@ -1,7 +1,7 @@
 const test = require('tape');
 const pg = require('pg');
 const { getConfig } = require('../../../config');
-const refreshDB = require('../../../db/scripts/refresh');
+const { refresh: refreshDB } = require('../../../db/scripts');
 const userCheckExists = require('../../../react-backend/database/queries/user_check_exists');
 
 const config = getConfig(process.env.NODE_ENV);
